Rename StringEnumEnumerableConverter to StringEnumReadOnlyListConverter
Migrate from JsonConverter<IEnumerable<StringEnum<T>>> to
JsonConverter<IReadOnlyList<StringEnum<T>>>, unblocking the last 4
IEnumerable properties to use IReadOnlyList.
