+++ src/Octokit.Webhooks/Models/App.cs
@@ -39,6 +39,6 @@ public sealed record App
     public AppPermissions? Permissions { get; init; }
 
     [JsonPropertyName("events")]
-    [JsonConverter(typeof(StringEnumEnumerableConverter<AppEvent>))]
-    public IEnumerable<StringEnum<AppEvent>>? Events { get; init; }
+    [JsonConverter(typeof(StringEnumReadOnlyListConverter<AppEvent>))]
+    public IReadOnlyList<StringEnum<AppEvent>>? Events { get; init; }
 }
