+++ src/Octokit.Webhooks/Models/MetaEvent/Hook.cs
@@ -16,8 +16,8 @@ public sealed record Hook
     public bool Active { get; init; }
 
     [JsonPropertyName("events")]
-    [JsonConverter(typeof(StringEnumEnumerableConverter<AppEvent>))]
-    public required IEnumerable<StringEnum<AppEvent>> Events { get; init; }
+    [JsonConverter(typeof(StringEnumReadOnlyListConverter<AppEvent>))]
+    public required IReadOnlyList<StringEnum<AppEvent>> Events { get; init; }
 
     [JsonPropertyName("config")]
     public required HookConfig Config { get; init; }
