+++ test/Octokit.Webhooks.Test/Converter/StringEnumReadOnlyListConverterTestsKnownData.cs
@@ -5,11 +5,10 @@ using Octokit.Webhooks.Extensions;
 using Octokit.Webhooks.Models.PingEvent;
 using Xunit;
 
-public class StringEnumEnumerableConverterTestsUnknownData : TheoryData<string, IEnumerable<StringEnum<HookType>>>
+public class StringEnumReadOnlyListConverterTestsKnownData : TheoryData<string, IReadOnlyList<StringEnum<HookType>>>
 {
-    public StringEnumEnumerableConverterTestsUnknownData() => this.Add(@"[""App"", ""Organization"", ""Unknown""]", [
+    public StringEnumReadOnlyListConverterTestsKnownData() => this.Add(@"[""App"", ""Organization""]", [
         new StringEnum<HookType>(HookType.App),
         new StringEnum<HookType>(HookType.Organization),
-        new StringEnum<HookType>("Unknown"),
     ]);
 }
